Fix for changed output of BB in GeoSteiner v5.0.1
This project was built for Geosteiner v3 or v3.1. With the most recent
version (5.0.1) the output of bb has had a breaking change from the
nomenclature of "steiner minimal" to "Euclidean SMT". This results in
the script including extra lines from the output that do not match the
expected structure.

- Replaced '(Steiner Minimal' in line 18 with 'Euclidean SMT'